Changed version from revision 8.1 to subversion 81
This is to prevent connecting to older versions of the MegaMod via Net Melee

+++ src/uqm/supermelee/netplay/packet.h
@@ -143,7 +143,7 @@ typedef struct {
 	struct {
 		uint8 major;
 		uint8 minor;
-		uint8 patch;
+		float patch;
 	} uqmVersion; /* Protocol version */
 	uint8 padding1;  /* Set to 0 */
 } Packet_Init;
